refactor data fetching to sagas
