+++ src/home/saga.ts
@@ -1,6 +1,16 @@
 import { fetchSentEscrowPayments } from 'src/escrow/actions'
 import { notificationsChannel } from 'src/firebase/firebase'
-import { Actions, refreshAllBalances, setLoading, updateNotifications } from 'src/home/actions'
+import {
+  Actions,
+  refreshAllBalances,
+  setLoading,
+  updateCleverTapInboxMessages,
+  updateNotifications,
+} from 'src/home/actions'
+import {
+  ExpectedCleverTapInboxMessage,
+  cleverTapInboxMessagesChannel,
+} from 'src/home/cleverTapInbox'
 import { IdToNotification } from 'src/home/reducers'
 import { fetchCurrentRate } from 'src/localCurrency/actions'
 import { shouldFetchCurrentRate } from 'src/localCurrency/selectors'
@@ -100,8 +110,30 @@ function* fetchNotifications() {
   }
 }
 
+export function* fetchCleverTapInboxMessages() {
+  const channel = yield* call(cleverTapInboxMessagesChannel)
+
+  try {
+    while (true) {
+      const notifications = (yield* take(channel)) as ExpectedCleverTapInboxMessage[]
+      yield* put(updateCleverTapInboxMessages(notifications))
+    }
+  } catch (error) {
+    Logger.error(
+      `${TAG}@updateCleverTapInboxMessages`,
+      'Failed to update CleverTap Inbox messages',
+      error
+    )
+  } finally {
+    if (yield* cancelled()) {
+      channel.close()
+    }
+  }
+}
+
 export function* homeSaga() {
   yield* spawn(watchRefreshBalances)
   yield* spawn(autoRefreshWatcher)
   yield* spawn(fetchNotifications)
+  yield* spawn(fetchCleverTapInboxMessages)
 }
